Switch to PolyMetriX Update smiles_to_graph, corrected torch.tensor errors W.I.P implementation of the GCN 	modified:   RDKit_testing_polymers.ipynb 	modified:   tools.py

+++ tools.py
@@ -16,7 +16,7 @@ def smiles_to_graph(smiles):
     for atom in mol.GetAtoms():
         atom_features = [atom.GetAtomicNum(), atom.GetDegree(), atom.GetFormalCharge(), atom.IsInRing(), atom.GetIsAromatic(), int(atom.GetHybridization())]
         nodes_features.append(atom_features)
-    nodes_features = torch.tensor(nodes_features, dtype=torch.long)
+    nodes_features = torch.tensor(nodes_features, dtype=torch.float)
 
     # Create the data object
     data = Data(x=nodes_features, edge_index=edge_index)
